feat: move image block to blocky-react
