+++ packages/blocky-react/src/blocks/index.ts
@@ -1,0 +1,1 @@
+export * from "./imageBlock";
