+++ packages/blocky-react/tsconfig.json
@@ -33,6 +33,7 @@
   },
   "include": [
     "src",
+    "src/blocks",
     "plugins"
   ],
   "references": [
